requirements.txt
pytz==2024.2
rfc5424-logging-handler==1.4.3
tzlocal==5.2
